flask/requirements.txt
random2==1.0.1
numpy==1.20.1
Babel==2.9.0
Jinja2==2.11.3
requests==2.25.1
validators==0.18.2
dash==1.20.0
dash_cytoscape==0.2.0
Werkzeug==1.0.1
WTForms==2.3.3
Flask==1.1.2
visdcc==0.0.40
pandas==1.2.2
dash_html_components==1.1.3
dash_core_components==1.16.0
Flask_WTF==0.14.3
fuzzywuzzy==0.18.0
dash_bootstrap_components==0.13.1
python-dotenv==0.19.2
gql==3.0.0a5
aiohttp==3.7.3

async-timeout==3.0.1
attrs==20.3.0
Brotli==1.0.9
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask-Compress==1.9.0
future==0.18.2
graphql-core==3.1.3
idna==2.10
install==1.3.4
itsdangerous==1.1.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
multidict==5.1.0
Pillow==8.1.0
plotly==4.14.3
promise==2.3
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytz==2021.1
retrying==1.3.3
Rx==1.6.1
scipy==1.6.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
websockets==8.1
yarl==1.6.3
Flask_Babel==2.0.0
